feature(document): create actions and reducers for creating documents  [Finishes #147341343]
