+++ client/app/src/constants/actionTypes.js
@@ -2,3 +2,5 @@ export const CREATE_USERS = 'CREATE_USERS';
 export const FETCH_USERS = 'FETCH_USERS';
 export const CREATE_DOCUMENT = 'CREATE_DOCUMENT';
 export const SET_LOGIN_USERS = 'SET_LOGIN_USERS';
+export const FETCH_USER_DOCUMENTS = 'FETCH_USER_DOCUMENTS';
+export const FETCH_ALL_DOCUMENTS = 'FETCH_ALL_DOCUMENTS';
